Rename bitwarden-json-types
